% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainKmeans.R
\name{myKmeans}
\alias{myKmeans}
\title{Cette fonction effectue la somme entre deux nombres}
\usage{
myKmeans(x, K, nbinit, nbCPU, useC)
}
\arguments{
\item{x}{numeric. Matrice des donnees (1 ligne par observation).}

\item{K}{numeric. Nombre de clusters.}

\item{nbinit}{numeric. Nombre d'initalisations aléatoires de l'algorithme.}

\item{nbCPU}{numeric. Nombre de CPU (si nbCPU>1, on parallelise).}

\item{useC}{boolean. Si TRUE, on appelle le code C++.}
}
\value{
list
}
\description{
Cette fonction effectue la somme entre deux nombres
}
\examples{
set.seed(123)
ech <- simuledata(100, 3)

set.seed(123)
T1 <- Sys.time()
res <- myKmeans(ech$x, 4, 200, 1, FALSE)
T2 <- Sys.time()
set.seed(123)
res2 <- myKmeans(ech$x, 4, 200, 2, TRUE)
T3 <- Sys.time()
all.equal(res, res2)
difftime(T3, T2)
difftime(T2, T1)

}
