% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedeDocumentation.R
\docType{package}
\name{packageKmeans-package}
\alias{packageKmeans-package}
\alias{packageKmeans}
\title{packageKmeans décrit en une ligne}
\description{
Ici je décris le but du package
}
\details{
\tabular{ll}{
Package: \tab packageKmeans\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-01-06\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\keyword{package}
