% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsFunctions.R
\name{simuledata}
\alias{simuledata}
\title{Cette fonction effectue la somme entre deux nombres}
\usage{
simuledata(n, delta)
}
\arguments{
\item{n}{numeric. Taille de l'echantillon.}

\item{delta}{numeric. Parametre qui controle la separation des classes}
}
\value{
list
}
\description{
Cette fonction effectue la somme entre deux nombres
}
\examples{
ech <- simuledata(50, 3)
}
